Fix type bugs
int to long type for chat_id
int to uint type for message_id
Added GetUpdateType, an extension method for Update.

+++ botAPI/Available Methods/getChat.cs
@@ -13,6 +13,6 @@ namespace TelegramAPI.Available_Methods
         /// <param name="T">BotClient</param>
         /// <param name="chat_id">Unique identifier for the target chat or username of the target supergroup (in the format @supergroupusername).</param>
         /// <returns>Chat Object.</returns>
-        public static Available_Types.Chat GetChat(this BotClient T, int chat_id) => T.RPC<Available_Types.Chat>("getChat", new JObject { new JProperty("chat_id", chat_id) });
+        public static Available_Types.Chat GetChat(this BotClient T, long chat_id) => T.RPC<Available_Types.Chat>("getChat", new JObject { new JProperty("chat_id", chat_id) });
     }
 }
